requirement.txt
# Core Django dependencies
Django==5.1.5
django-environ==0.12.0
django-filter==24.3

# Database and storage
pillow==11.1.0
whitenoise==6.8.2

# Core Python dependencies
asgiref==3.8.1
sqlparse==0.5.3
six==1.17.0

# Development tools (optional)
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Supporting packages
click==8.1.8
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycodestyle==2.12.1
pyflakes==3.2.0